Add timer method and trigger updateFutureDateTime
This timer intends to update the ftxLockTime field in sync with system time until the user changes it. 
There is minor skew in the calculated seconds that is offset by making the referenced seconds into a float. Then converting back the rounded number compared to default 86400 seconds (1 day as default)